+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.11.2](https://github.com/webpod/zurk/compare/v0.11.1...v0.11.2) (2025-04-01)
+
+### Fixes & improvements
+* fix: export `VoidStream` ([626d865](https://github.com/webpod/zurk/commit/626d86561774f1bafda21eb750ee42ed73c39ee1))
+
+## [0.11.1](https://github.com/webpod/zurk/compare/v0.11.0...v0.11.1) (2025-03-31)
+
+### Fixes & improvements
+* refactor: use .ts ext in sources ([2504eab](https://github.com/webpod/zurk/commit/2504eabe20310f8287efcdfeb4a03f0ea4fe2b00))
+
 ## [0.11.0](https://github.com/webpod/zurk/compare/v0.10.3...v0.11.0) (2025-03-27)
 
 ### Features